Add files via upload
In this project, We had to develop efficient algorithms to solve the problem of finding all possible strings of length L which have weight W and hamming distance of D with all other strings. Applications of this problem ranges from communications to coding theory. I've implemented my solver with SA algorithm and with two different methods, first trying to make all strings and then attempting to edit strings so as to make them in appropriate shape, in another hand then second method focuses on starting from two string and tries to add more strings as much it can.